Add additional key actions for Row editing
Since the rows auto-save on blur, allow a way to cancel the edit by pressing escape.

Also, hitting "tab" would natively skip to the next input field on the page. Change that to jump to editing the next row in the list. We have to use events to trigger this, since a given Row View has no link to any other Row View, but the Row knows its own Model, which has a link to its Collection. So, we find another Model in the collection and trigger an event on it, which is heard by its View, wherever that is.