Use local storage for recent visits
Use HTML5's local storage ability to save a list of recently-visited Grimoires (public_key), and to save the admin_key of any Grimoire visited in which the admin key is known. Since there's no logins for this service, the browser accessing the admin URL indicates ownership, and therefore we can add to our "keychain" of known keys.