Remove reliance on text input 'blur' events
Saving the element when an inline text box gets a 'blur' event makes it impossible to do things like an auto-complete popup option. This changes the system to fire an event when something else in the DOM is clicked, and doesn't act upon it, so it bubbles up to the body.